tensorflow-lite-applications: introduce C/C++ image classification app
This application demonstrate a computer vision use case for image
classification where frames are grabbed from a camera input
(/dev/videox) and analyzed by a neural network model interpreted by
TensorFlow Lite framework.

Gstreamer pipeline is used to stream camera frames (using v4l2src), to
display a preview (using waylandsink) and to execute neural network
inference (using appsink).

The result of the inference is displayed on the preview. The overlay is
done using GTK widget with cairo.

This combination is quite simple and efficient in term of CPU
consumption.

Signed-off-by: Vincent Abriou <vincent.abriou@st.com>
Change-Id: Ia4a70af832443b280b2440e443342c33e7dded97
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/152102