tensorflow-lite-applications: introduce C/C++ object detection app
This application demonstrate a computer vision use case for object
detection where frames are grabbed from a camera input (/dev/videox) and
analyzed by a neural network model interpreted by TensorFlow Lite
framework.

Gstreamer pipeline is used to stream camera frames (using v4l2src), to
display a preview (using waylandsink) and to execute neural network
inference (using appsink).

The result of the inference is displayed on the preview. The overlay is
done using GTK widget with cairo.

This combination is quite simple and efficient in term of CPU
consumption.

Signed-off-by: Vincent Abriou <vincent.abriou@st.com>
Change-Id: I6c0545f84117c0cfcd8f0d56dab82af8893cd3b9
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/152103