tensorflow-lite: remove bashism patch and use bash in the bitbake recipe
The "remove bashism" patch is not accepted for upstream.
So to avoid the lite/tools/pip_package/build_pip_package.sh script failure
in a host machine that by default have the dash shell enabled, we force to
execute the build_pip_package.sh script with bash in the bitbake recipe.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>
Change-Id: I0ca4820a5ad5f32c40805d086684195a70ecff6c
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/158164