create packages to install only NN models
Remove the installation of the NN models from the application recipes.
Thus, models are installed once for different purpose like ArmNN and TFLite
application examples.
The installation directory of the NN models should also be changed to be
used by different applications and as a consequence, the installation
directory for TFLite application must be change too.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>