tflite-edgetpu-app: Introduce image classification Python demo application
This application demonstrate a computer vision use case for image classification
where frames are grabbed from a camera input (/dev/videox) and
analyzed by a neural network model interpreted by TensorFlow Lite
framework. It's using the experimental_delegates API to load the libedgetpu.so

The GTK Library is used to display the results of the inference.

The processes are running sequentially and in IDLE mode. This application
is allowing enormous gains in terms of inference time by calling the
AI Coral accelerator.

Signed-off-by: Othmane <othmane.ahlzouaoui@st.com>
Change-Id: I0f5a83c46a3057b2ccbf3c62a135c18e5b38facc
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/169844
Tested-by: Vincent ABRIOU <vincent.abriou@st.com>
Reviewed-by: Vincent ABRIOU <vincent.abriou@st.com>