tflite-edgetpu-apps: Introduce object detection Python demo application
This application demonstrate a computer vision use case for object
detection where frames are grabbed from a camera input (/dev/videox) and
analyzed by a neural network model interpreted by TensorFlow Lite
framework. It's using the experimental_delegates API to load the libedgetpu.so

The GTK Library is used to display the results of the inference.

The processes are running sequentially and in IDLE mode. This application
is allowing enormous gains in terms of inference time by calling the
AI Coral accelerator.

Signed-off-by: Othmane <othmane.ahlzouaoui@st.com>
Change-Id: Ifb8215218a33e12e61645347509cc85e34a036a0
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/169846
Tested-by: Vincent ABRIOU <vincent.abriou@st.com>
Reviewed-by: Vincent ABRIOU <vincent.abriou@st.com>