tflite-cv-apps: face reco application (BlazeFace + FaceNet)
The face reco application is based on 2 neural networks:
- BlazeFace model for face detection with landmarks
- A retrained FaceNet model for face recognition

The output of the BlazeFace NN is compute to align the detected faces
that are provided to the second FaceNet NN for face recognition.

To enroll a new face, just touch (or click) the unknown face and enter
the name through the virtual keyboard. The face is then added to database.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>