tflite-cv-apps: face recognition application
The face recognition application is based on 2 TensorFlow Lite neural
networks implemented within 2 libraries:
- the libfacedetect_tfl.a library for the face detection
- the libfacereco_tfl.a library for for face recognition

The output of the face detection neural network is computed and provided
to the second neural network to execute the face recognition on the
detected face.

To enroll a new face, just touch (or click) the unknown face and enter
the name through the virtual keyboard. The face is then added to database.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>
Change-Id: I2e797c37c2c38b368df35543d9b573a493dbe1f5
Reviewed-on: https://gerrit.st.com/c/stm32mpuapp/meta/meta-st-stm32mpu-ai/+/188632