tflite-cv-apps: facereco: replace number_of_faces parameter by a boolean multiple_faces parameter
The way the number_of_faces parameter is handle is wrong and actually
cannot be used to speedup face detection.
Decision has been taken to replace the number_of_faces parameter by a
single boolean parameter name mutliple_faces to selected either single face
detection (if the parameter is set to "false") or multiple faces detection
(if this parameter is set to "true").

related to: bug #847212

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>
Change-Id: I3e6479767a09bd11be1358f8750278a73a2df3cb