tflite-cv-apps : Object detection : Fixing max detection issues
C/C++ applications :

Modification of the object detection application and wrapper to allocate
dynamically the size of the result vector to handle each possible number
of max-detection object.

The detection results are no longer stored in a structure composed of
fixed-size arrays, they are stored in a vector of ObjDetect_Results struct
which the size is dynamically allocated depending on the memory demand.

For each frame a vector is build, the size of the vector depend of
the Max_detection number of the model used. Every object detected is
represented by 6 values (ID, Score, Y0, X0, Y1, X1).

The results structure is then composed of a vector of objects detected
and a float value representing the inference time.

Python applications :

For the python applications the way to fix max detection issues is
different from the c/c++ way due to the multiprocess method used.

For now the maximum detection value representing the number of objects
detected in one frame is an argument of our python applications.
The default value remains 10.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>