tflite-cv-apps : fix the average inference time in validation mode
This patch add a filter on the first 5 frames computed which remove their
inference times from the global average inference time returned in validation
mode using camera. This is necessary because of the functioning of the
gstreamer plugin used to recover the inference time. The validation mode is
now testing 100 frames instead of 25 before to be more accurate and smooth
out the values.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>