recipe-devtools: cmake-native_3.16.5 support git clone shallow with specified SHA1 ID
TensorFlow-Lite 2.7.0 is now using cmake as build system. Thusly
TensorFlow-Lite dependencies are downloaded using cmake via git clone.

Depending on the git tool configuration the cmake git clone of gitlab
repositories may fail with the following message: "The TLS connection was
non-properly terminated". This is the case when git tool is using gnuTLS
(no issue have been reported with git tool using OpenSSL).
To overcome this problem, the solution consist in configuring cmake to
clone the repository using GIT_SHALLOW option that enable "--depth 1"
cloning option. But the cmake GIT_SHALLOW implementation is not able to
clone only the specified SHA1 commit ID and the TensorFlow-Lite download
dependencies step is failing.

This patch provide a solution to perform GIT_SHALLOW on a SHA1 commit ID
and allow to speed up the overall cmake download dependencies mechanism.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>