tflite-cv-apps : python : image classification : update app
This patch modify the entire GUI of the python image classification
application, using GTK, Gstreamer and the wayland window system.
This new method allow more flexibility and control on the items displayed
on the screen.

The video stream mechanism has been modified. OpenCV videocapture is
no more used to recover and display the video stream. Now gstreamer
and the gtksink plugin are used to display the stream in a Gtk Widget.

The multiprocessing is now more efficient in case of a platform with more
than one cpu, the video stream take one core and the others are used by
the tflite interpreter for the inference.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>