onnx-cv-apps: onnxruntime application for object detection with C++
This patch introduces the onnxruntime application for object detection
with C++ .The necessary header files were installed in the
onnxruntime_git.bb .the src folder in onnxrt-cv-apps contains the files
that were added : the .sh files for launching and testing ,the .css files
for widgets ,the Makefile to configure c++ options and the .yaml file to
link with demo launcher .
the objdetect_onnx_gst_gtk.cc file is used to configure inputs and
display and calls the wrapper.hpp file which is used for inference.
the Onnxruntime icon was added in the ressources folder
the onnx-cv-apps-object-detect-c++_3.0.0.bb contains the bitbake recipe
to build this application.

Signed-off-by: youssef KHEMAKHEM <youssef.khemakhem@st.com>