onnx-cv-apps: onnxruntime application for object detection with python
This patch introduces the onnxruntime application for object detection
with python. The python folder in onnxrt-cv-apps contains the files
that were added : the .sh files for launching and testing, the .css files
for widgets and the .yaml file to link with demo launcher.
The objdetect_onnx.py file is used to configure inputs and display
and launch inference.
The onnx-cv-apps-object-detect-python_3.0.0.bb contains the bitbake
recipe to build this application.

Signed-off-by: youssef KHEMAKHEM <youssef.khemakhem@st.com>