frameworks: introduce stai_mpu AI unified API for STM32MPUs
The stai_mpu API is a cross STM32MPUs platforms neural network
inferencing API with a flexible interface to run several deep learning
models formats such as TensorFlow™ Lite[1], ONNX™[2] and Network Binary
Graph (NBG).

It provides support for C++ and Python language and is integrated as
part of the X-LINUX-AI package.

Signed-off-by: Vincent ABRIOU <vincent.abriou@st.com>