applications: add face recognition out-of-the box application
Add a stai_mpu out-of-the box application for face recognition.
This application is dedicated to run on STM32MP2x board and
relies on the DCMIPP dual pipeline camera.
The models used for the application are open source models:
-blazeface (128x128) for face detection
-facenet512 (160x160) for face recognition
This application demonstrate a simple usecase without any anti-spoofing
or security mechanism.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>