x-linux-ai-benchmark: add hw accel features for tflite and onnx frameworks
With this new version of the benchmark, for STM32MP2X platform, when a
tflite or an onnx model is provided the benchmark will automatically
try to run it using the GPU/NPU with tflite vx-delegate and onnx vsinpu
execution provider. If the benchmark fails with hardware acceleration
a second run will be done on CPU only.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>