st-image-ai: converge on a unique solution for all STM32MPU
To unify the st-image-ai for hardware that support AI hardware acceleration
and hardware that doesn't, an unique st-image-ai is now available, by
default for MP1X boards the image only support CPU inference, for STM32MP2X the default
image is using NPU acceleration. This functioning can be modify directly
in the layer.conf.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>